Refactor RadiusExample to Example tsx file
